Telegram_Bot/requirements.txt
langchain
langchain-google-genai
python-telegram-bot==20.7
google-generativeai
python-dotenv
requests
